fix: Pick up correct compose file in `_docker_compose`
The "internal" `_docker_compose` function is modified to take an
optional 1st argument.  If this argument is the string `remote`, then
this argument is discarded and the `docker-compose-remote.yml` file
will be used in addition to the ordinary `docker-compose.yml` file.

If the 1st argument is the string `local`, then this argument is
discarded and the `docker-compose-local.yml` file is used together with
`docker-compose.yml`.

If the 1st argument is any other string, or if it is missing, the 1st
argument is retained and the two files `docker-compose-local.yml` and
`docker-compose.yml` are used.