fix: Source the correct environment file in `_setup_environment`
Get rid of the `LOCALENV` variable as it's no longer required to keep
state.

Don't explicitly call `_setup-environment` before calling
`generate-data` as `generate-data` already calls that function.

Make `deploy-stack` set `CHOSEN_ENV` to the correct environment filename
depending on its first argument (if that argument is available,
otherwise act as if it was `local`).

In `_setup-environment`, don't modify `CHOSEN_ENV` if it is already set
(e.g. by `deploy-stack`).