Validate API input to prevent injection (#1366)
* Validate API input to prevent injection

The changes involve a range of fixes and improvements to various APIs, with a particular focus on preventing potential cypher injection attacks. This is achieved by validating user input and removing unnecessary logging in parts of the code that could be susceptible to such attacks. API tests were added to make sure that malicious chars are not accepted in the neo4j related endpoints. Additionally, we addressed a CodeQL warning by implementing a check to ensure that a parameter is not an array.

---------

Co-authored-by: Per Johnsson <per.johnsson@nbis.se>
Co-authored-by: jonandernovella <novella.uu@gmail.com>